Give command allow string "max"
Allow the string "max" in place of an integer for the number of items wanted. This will allow the user to type for example; give max snowballs, to give all the possible snowballs the server can without having to be presented with a message saying how many can be spawned, and then typing the command once more with the new value.