allowed give to give max
Give will now give Tools.AvailableItemSlots instead of what the user entered in the event the user entered more than Tools.AvailableItemSlots. This allows the command to go ahead and complete aith the max number of item slots available instead of just throwing an error and cancelling the command requiring the user to type it again with the number displayed.