disabled displaying message for no help text
Disabled sending message for when there is no helpText set for a
command.
This prevents the message from appearing even when there was a
description. There should be some checks done on all of the variables
set in entry.cs and a message displayed informing the user only if none
are set.